Update standardizeFromModel.R
If the 'shape' part of the model is 'none', the part of the 'cor' matrix containing the specified row and column analyte sets will be added to the output matrix as-is. This allows two things: (1) preparing the standardized matrix ahead of time, and (2) using this code to construct a large correlation matrix from smaller pieces without performing standardization.